requirements.txt
uvicorn
FastAPI
starlette
passlib
pydantic
pymysql
python-jose
python-multipart
cryptography